// src/API.ts
export type EventTemplate = {
  __typename: "EventTemplate",
  userId?: string | null,
  eventName: string,
  description?: string | null,
  duration: number,
  id: string,
};